perf: 🚀  optimization multiple tab list search, add logging alert rule components, optimization wor
