+++ src/mixins/base_filter.js
@@ -2,8 +2,12 @@ const basefilter = {
   inject: ['reload'],
   methods: {
     async m_filter_list(params, reload = true) {
+      // 处理多tab列表
+      const tab = this.$route.query.tab || null
       Object.assign(this.params, params)
-      await this.$router.replace({ name: this.$route.name, query: params })
+      await this.$router.replace({
+        name: this.$route.name, query: Object.assign(params, { tab: tab }),
+      })
       if (reload) {
         this.reload()
       }
