+++ src/router/admin_observe.js
@@ -203,7 +203,7 @@ export const adminObserve = [
           {
             path: `overview`,
             name: `${prefix}-observe-monitor-overview`,
-            component: () => import('@/views/observe/monitor/overview'),
+            component: () => import('@/views/observe/alert/dashboard'),
             meta: {
               requireAuth: true,
               title: '概览',
@@ -218,7 +218,7 @@ export const adminObserve = [
           {
             path: `living`,
             name: `${prefix}-observe-monitor-alert`,
-            component: () => import('@/views/observe/monitor/alert'),
+            component: () => import('@/views/observe/alert/living'),
             meta: {
               requireAuth: true,
               title: '实时告警',
@@ -233,7 +233,7 @@ export const adminObserve = [
           {
             path: `history`,
             name: `${prefix}-observe-alert-history`,
-            component: () => import('@/views/observe/monitor/history'),
+            component: () => import('@/views/observe/alert/history'),
             meta: {
               requireAuth: true,
               title: '告警历史',
@@ -248,7 +248,7 @@ export const adminObserve = [
           {
             path: `blacklist`,
             name: `${prefix}-observe-alert-blacklist`,
-            component: () => import('@/views/observe/monitor/blacklist'),
+            component: () => import('@/views/observe/alert/blacklist'),
             meta: {
               requireAuth: true,
               title: '告警黑名单',
