+++ src/views/microservice/workload/index.vue
@@ -26,6 +26,7 @@
           </v-tab>
         </v-tabs>
         <v-data-table
+          class="kubegems__table-row-pointer"
           disable-sort
           :headers="headers"
           :items="items"
@@ -357,21 +358,28 @@ export default {
     EnvironmentFilter,
   },
   mixins: [BasePermission, BaseFilter, BaseResource, BaseSelect],
-  data: () => ({
-    tab: 0,
-    tabItems: [
-      { text: '无状态服务', value: 'Deployment' },
-      { text: '有状态服务', value: 'StatefulSet' },
-    ],
-    items: [],
-    pageCount: 0,
-    params: {
-      page: 1,
-      size: 10,
-    },
-    filters: [{ text: '负载名称', value: 'search', items: [] }],
-    podItems: [],
-  }),
+  data() {
+    this.tabMap = {
+      deployment: 0,
+      statefulset: 1,
+    }
+
+    return {
+      tab: this.tabMap[this.$route.query.tab] || 0,
+      tabItems: [
+        { text: '无状态服务', value: 'Deployment', tab: 'deployment' },
+        { text: '有状态服务', value: 'StatefulSet', tab: 'statefulset' },
+      ],
+      items: [],
+      pageCount: 0,
+      params: {
+        page: 1,
+        size: 10,
+      },
+      filters: [{ text: '负载名称', value: 'search', items: [] }],
+      podItems: [],
+    }
+  },
   computed: {
     ...mapState(['JWT', 'MessageStreamWS', 'EnvironmentFilter']),
     ...mapGetters(['VirtualSpace', 'Environment']),
@@ -438,6 +446,14 @@ export default {
       },
       deep: true,
       immediate: true,
+    },
+    tab() {
+      this.$router.replace({
+        query: {
+          ...this.$route.query,
+          tab: this.tabItems[this.tab].tab,
+        },
+      })
     },
   },
   methods: {
