+++ src/views/observe/alert/blacklist/index.vue
@@ -15,6 +15,7 @@
       </div>
 
       <v-data-table
+        class="kubegems__table-row-pointer"
         :headers="headers"
         :items="items"
         :page.sync="params.page"
