+++ src/views/observe/alert/dashboard/components/AlertCategoryBar.vue
