+++ src/views/observe/alert/living/index.vue
