+++ src/views/observe/log/config/alert_rule/index.vue
@@ -83,6 +83,7 @@
       </v-card-title>
       <v-card-text class="px-0">
         <v-data-table
+          class="kubegems__table-row-pointer"
           disable-sort
           disable-filtering
           :headers="headers"
@@ -295,12 +296,12 @@
     </v-card>
     <AddAlertRule
       ref="addAlertRule"
-      mode="log"
+      mode="logging"
       @refresh="alertRuleList"
     />
     <UpdateAlertRule
       ref="updateAlertRule"
-      mode="log"
+      mode="logging"
       @refresh="alertRuleList"
     />
   </v-container>
@@ -537,7 +538,7 @@ export default {
     },
     onPageSizeChange(size) {
       this.params.page = 1
-      this.params.size = size
+      this.itemsPerPage = size
     },
     onPageIndexChange(page) {
       this.params.page = page
