+++ src/views/resource/environment/index.vue
@@ -9,8 +9,10 @@
       <v-card-title class="py-4">
         <BaseFilter
           :filters="filters"
+          :reload="false"
           :default="{ items: [], text: '环境名称', value: 'search' }"
           @refresh="m_filter_list"
+          @filter="customFilter"
         />
         <v-sheet class="text-subtitle-2 ml-4">租户</v-sheet>
         <v-sheet width="350">
@@ -53,9 +55,11 @@
           disable-sort
           :headers="headers"
           :items="items"
-          :items-per-page="500"
+          :page.sync="page"
+          :items-per-page="itemsPerPage"
           no-data-text="暂无数据"
           hide-default-footer
+          @page-count="pageCount = $event"
         >
           <template #[`item.environmentName`]="{ item }">
             {{ item.EnvironmentName }}
@@ -183,6 +187,17 @@
             </v-menu>
           </template>
         </v-data-table>
+
+        <BasePagination
+          v-if="pageCount >= 1"
+          v-model="page"
+          :front-page="true"
+          :page-count="pageCount"
+          :size="itemsPerPage"
+          @loaddata="environmentTenantResourceQuota(tenant)"
+          @changesize="onPageSizeChange"
+          @changepage="onPageIndexChange"
+        />
       </v-card-text>
     </v-card>
 
@@ -203,6 +218,7 @@ import BasePermission from '@/mixins/permission'
 import BaseFilter from '@/mixins/base_filter'
 import BaseTable from '@/mixins/table'
 import { sizeOfStorage, sizeOfCpu } from '@/utils/helpers'
+import { deepCopy } from '@/utils/helpers'
 
 export default {
   name: 'Environment',
@@ -213,8 +229,12 @@ export default {
   inject: ['reload'],
   data: () => ({
     items: [],
+    itemsCopy: [],
     tenant: -1,
     params: {},
+    page: 1,
+    pageCount: 0,
+    itemsPerPage: 10,
     filters: [{ text: '环境名称', value: 'search', items: [] }],
   }),
   computed: {
@@ -262,6 +282,21 @@ export default {
     }
   },
   methods: {
+    customFilter() {
+      if (this.$route.query.search && this.$route.query.search.length > 0) {
+        this.items = this.itemsCopy.filter((item) => {
+          return (
+            item.EnvironmentName &&
+            item.EnvironmentName
+              .toLocaleLowerCase()
+              .indexOf(this.$route.query.search.toLocaleLowerCase()) > -1
+          )
+        })
+      } else {
+        this.items = this.itemsCopy
+      }
+      // this.m_table_generateSelectResource()
+    },
     async environmentTenantResourceQuota(tenantid) {
       const data = await getEnvironmentTenantResourceQuota(tenantid, this.params)
       this.items = data
@@ -308,6 +343,8 @@ export default {
         e.StoragePercentage =
           e.Storage > 0 ? ((e.UsedStorage / e.Storage) * 100).toFixed(1) : 0
       })
+      this.itemsCopy = deepCopy(this.items)
+      if (this.$route.query.search) this.customFilter()
     },
     onTenantSelectChange() {
       if (this.tenant) this.environmentTenantResourceQuota(this.tenant)
@@ -348,6 +385,13 @@ export default {
     },
     onTenantSelectFocus() {
       this.m_select_tenantSelectData()
+    },
+    onPageSizeChange(size) {
+      this.page = 1
+      this.itemsPerPage = size
+    },
+    onPageIndexChange(page) {
+      this.page = page
     },
   },
 }
