+++ src/views/resource/workload/components/WorkloadBaseForm/ContainerImage/container_section/ContainerPort/index.vue
@@ -69,82 +69,13 @@
         </v-card-actions>
       </v-card>
     </v-expand-transition>
-    <v-sheet
-      v-for="(item, index) in containerCopy && containerCopy.ports
-        ? containerCopy.ports
-        : []"
-      :key="index"
-      class="grey lighten-4 rounded ma-2"
-    >
-      <v-list-item two-line>
-        <v-list-item-content class="py-2">
-          <v-list-item-subtitle class="text-body-2 py-0">
-            <v-list-item
-              two-line
-              class="float-left pa-0 kubegems__three-width"
-            >
-              <v-list-item-content class="py-0">
-                <v-list-item-title class="text-subtitle-2 py-1">
-                  {{ item.name }}
-                </v-list-item-title>
-                <v-list-item-subtitle class="text-body-2 py-1">
-                  名称
-                </v-list-item-subtitle>
-              </v-list-item-content>
-            </v-list-item>
-            <v-list-item
-              two-line
-              class="float-left pa-0 kubegems__three-width"
-            >
-              <v-list-item-content class="py-0">
-                <v-list-item-title class="text-subtitle-2 py-1">
-                  {{ item.protocol }}
-                </v-list-item-title>
-                <v-list-item-subtitle class="text-body-2 py-1">
-                  协议
-                </v-list-item-subtitle>
-              </v-list-item-content>
-            </v-list-item>
-            <v-list-item
-              two-line
-              class="float-left pa-0 kubegems__three-width"
-            >
-              <v-list-item-content class="py-0">
-                <v-list-item-title class="text-subtitle-2 py-1">
-                  {{ item.containerPort }}
-                </v-list-item-title>
-                <v-list-item-subtitle class="text-body-2 py-1">
-                  容器端口
-                </v-list-item-subtitle>
-              </v-list-item-content>
-            </v-list-item>
-          </v-list-item-subtitle>
-          <div class="kubegems__clear-float" />
-        </v-list-item-content>
-        <v-btn
-          dark
-          text
-          fab
-          right
-          x-small
-          color="primary"
-          @click="updateData(index)"
-        >
-          <v-icon>mdi-pencil</v-icon>
-        </v-btn>
-        <v-btn
-          dark
-          text
-          fab
-          right
-          x-small
-          color="error"
-          @click="removeData(index)"
-        >
-          <v-icon>mdi-delete</v-icon>
-        </v-btn>
-      </v-list-item>
-    </v-sheet>
+
+    <Port
+      :container-copy="containerCopy"
+      @updateData="updateData"
+      @removeData="removeData"
+    />
+
     <v-flex class="grey lighten-4 rounded-0 ma-2">
       <v-list-item two-line>
         <v-list-item-content class="py-2">
@@ -170,11 +101,15 @@
 </template>
 
 <script>
+import Port from './Port'
 import { deepCopy } from '@/utils/helpers'
 import { required, port } from '@/utils/rules'
 
 export default {
   name: 'ContainerPort',
+  components: {
+    Port,
+  },
   props: {
     container: {
       type: Object,
